Implementation of lepton mva's
This commit implements both
- the ttH lepton mva (HIG-18-019)
- the ghent lepton mva (TOP-18-008, TOP-18-009)

Possibly more lepton mva's, which include trainings for 2018 will be
implemented soon. The code should be flexible enough to deal with the
changes in variables which very likely will occur again.

I think the older implementation in electronsExtrasSUSY_cff.py is broken, but
it is not needed anymore with this new (easier) implementation. I only
leave the SUSY implementation in for now as a reference as it contains
some high-level variables (e.g. multi iso) which are not yet
re-implemented.

This new implementation is also not storing workingpoints themselves
(e.g. "passingLeptonMvaVeryTight" or similar), but only saves the values
and assumes analyzers can put the cut themselves in the fitter package.
In this we avoid adding many more booleans to the tree for slightly
different workingpoints.