Refactoring the probe modules
This commit has no impact on the tree content, except for a few variable
names who got a more consistent naming (e.g. the photon Fall17V2
workingpoints are reffered to as 94XV2 instead of 100XV2, given that the
former is a more official.

This commit tries to avoid the code duplication which was present in
this package. By refactoring we try to make the implementation of new
workingpoints easier, adding a new workingpoint new requires only one
new line in egmElectronIDModules_cff.py or egmPhotonIDModules_cff.py and
will automatically result in a new "passing<WP>" boolean.