Fix/improve compareTrees script
Cleaned-up the compareTrees script to simply compare using the branch
names. This is important because we wouldn't be CMS if someone in the
cmssw/PhysicsTools/TagAndProbe/src/BaseTreeFiller.cc fucked up the
branch and leaf names to be inconsistent. The old version of this script
couldn't compare such fucked-up branches.